refactor : modify RQ-003 mvc로 분리
