+++ src/main/java/baseball/Model/Referee.java
@@ -1,36 +1,7 @@
 package baseball.Model;
 
 public class Referee {
-    private int[] getQueryDecoding(String queryString) {
-        int asciiZero = 48;
-        int[] query = new int[3];
-        for(int i = 0; i < queryString.length(); ++i)
-            query[i] = queryString.charAt(i) - asciiZero;
-        return query;
-    }
+    private String exceptionString = "1부터 9까지 서로 다른 수로 이루어진 3자리의 수를 입력해주세요!";
 
-    public String judge(String queryString, Computer computer) {
-        int strikeCount = 0;
-        int ballCount = 0;
-        int[] query = getQueryDecoding(queryString);
-        int[] answer = computer.getAnswer();
-        boolean[] pickedMap = computer.getPickedMap();
-        StringBuilder sb = new StringBuilder();
-
-        for(int i = 0; i < 3; ++i) { 
-            if(pickedMap[query[i]]) {
-                if(query[i] == answer[i]) ++strikeCount;
-                else ++ballCount;
-            }
-        }
-
-        if(strikeCount == 0 && ballCount == 0)
-            return "낫싱";
-        if(ballCount > 0)
-            sb.append(ballCount).append("볼 ");
-        if(strikeCount > 0)
-            sb.append(strikeCount).append("스트라이크 ");
-
-        return sb.toString();
-    }
+    public String getExceptionString() {return exceptionString;}
 }
