+++ src/main/java/baseball/View/RefereeTerminalOutputView.java
@@ -1,5 +1,7 @@
 package baseball.View;
 
 public class RefereeTerminalOutputView {
-    
+    public void displayJudge(String judgeResult) {
+        System.out.println(judgeResult);
+    }
 }
